Introduce monorepo structure.
