+++ .tool-versions
@@ -1,1 +1,2 @@
 nodejs 18.12.1
+yarn 1.22.19
