+++ package.json
@@ -1,34 +1,21 @@
 {
-  "name": "@ramseyinhouse/steppp",
-  "version": "0.0.4",
-  "main": "dist/steppp.umd.js",
-  "module": "dist/steppp.es.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/"
-  ],
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build && tsc",
-    "serve": "vite preview",
-    "test": "tsc && vitest run",
-    "prettier": "prettier --write \"**/*.{js,ts,css,md}\"",
-    "prepare": "husky install"
-  },
-  "devDependencies": {
-    "@testing-library/dom": "^8.19.1",
-    "@types/node": "^18.11.18",
-    "@types/web-animations-js": "^2.2.12",
-    "husky": ">=8.0.3",
-    "jsdom": "^20.0.3",
-    "lint-staged": ">=13.1.0",
-    "prettier": "2.8.1",
-    "terser": "5.16.1",
-    "typescript": "^4.9.4",
-    "vite": "^4.0.4",
-    "vitest": "^0.26.3"
-  },
-  "lint-staged": {
-    "*.{js,ts,css,md}": "prettier --write"
-  }
+    "name": "@ramseyinhouse/steppp-monorepo",
+    "private": true,
+    "scripts": {
+        "build": "lerna run build",
+        "test": "lerna run test",
+        "graph": "nx graph",
+        "format": "prettier --write \"**/*.{js,ts,tsx,jsx,css,md,html}\"",
+        "publish": "lerna publish"
+    },
+    "lint-staged": {
+        "*.{js,ts,css,md}": "prettier --write"
+    },
+    "devDependencies": {
+        "lerna": "^6.4.1",
+        "prettier": "2.8.3"
+    },
+    "workspaces": [
+        "packages/*"
+    ]
 }
