+++ packages/steppp/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "outDir": "dist",
+    },
+    "include": [
+        "./src",
+        "integration/index.test.js"
+    ]
+}
